requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
Werkzeug==2.0.1 